gateway: harden path prefix
Only allow paths, so we stay local to this gateway,
and don't get tricked into linking to a different host,
or even redirecting to one.

I can't come up with a concrete attack right off the bat,
but who knows how things are going to change.
Not restricting it would be silly.

License: MIT
Signed-off-by: Lars Gierth <larsg@systemli.org>